# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend for the all rated extreme demon list",
        "title": "Aredl API",
        "contact": {
            "name": "Aredl",
            "url": "https://aredl.net/"
        },
        "version": "1.0"
    },
    "host": "api.aredl.net",
    "basePath": "/api",
    "paths": {
...
        "/aredl/names": {
            "get": {
                "description": "Gives a map of important users grouped by their role. This also includes aredl plus members",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Important users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/aredl.NameUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/mod/users/{id}/merge": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Directly merges two users\nRequires user permission: user_merge",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Merge two users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "primary user that the data gets merged into",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secondary user that gets deleted",
                        "name": "secondary_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/name-change-requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rejects a name change request from a user\nRequires user permission: name_change_review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Reject name change request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/create-placeholder": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a placeholder user for users that are not registered on the list yet\nRequires user permission: create_placeholder",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Create a placeholder user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "display name",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/global.CreatePlaceholderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/ban": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Bans a user and removes them from the leaderboard\nRequires user permission: user_ban\nAdditionally the user needs to be able to affect the user with their permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Ban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "internal user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/role": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Promote or demote a user\nRequires user permission: user_change_role\nAdditionally the user needs to be able to affect the user with their permission and give the user the new role",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Change user role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "internal user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "new roles",
                        "name": "roles",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "aredl.NameUser": {
            "type": "object",
            "properties": {
                "global_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "global.CreatePlaceholderResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "util.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Perform actions as a user. It is also used to access endpoints that require user permissions.",
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}